Add additional supported architectures to metadata
Previously, the library's metadata only specified compatibility with the `mbed` architecture. This is the architecture used by the "[DEPRECATED - Please install standalone packages] Arduino Mbed OS Boards" platform. As you might guess from the name, this platform is deprecated. Since it is clearly marked as such and has been so for some time, it is likely the majority of users have switched to the more specific "Arduino Mbed OS ___ Boards" platforms.

Since compatibility with those architectures was not specified in the metadata, the following impacts resulted on users who had one of the boards of the non-deprecated platforms selected:

A warning was displayed in the compilation output. For example

WARNING: library Arduino_PF1550-master claims to run on mbed architecture(s) and may be incompatible with your current board which runs on mbed_portenta architecture(s).

The library's examples were placed under the File > Examples > INCOMPATIBLE menu of Arduino IDE 1.x (example categorization by compatibility not implemented in Arduino IDE 2.x at this time), and the equivalent in the `arduino-cli lib examples` output.

The library was not listed in the `arduino-cli lib list --fqbn` output.