Fix Portenta C33 job configuration in "Compile Examples" CI workflow
Previously the "Compile Examples" GitHub Actions workflow was configured to install the boards platform dependency of
the "Portenta C33" board via a package index published for the purposes of beta testing during the initial development
of the "Arduino Renesas Portenta Boards" platform. That package index was later deleted, which caused the job to fail:

```
Downloading index: package_renesas_index.json 0 B / ?    0.00%
Downloading index: package_renesas_index.json Server responded with: 404 Not Found
Some indexes could not be updated.
```

The "Arduino Renesas Portenta Boards" platform is now part of the primary package index so the job can be restored to a
working state. Since that time, the board ID for the "Portenta C33" board was changed to `portenta_c33` so the FQBN of
the board is also updated here.